一部をリファクタリング

+++ slokara/slokara/Dormitory/Parameter/ParameterModel.swift
@@ -2,7 +2,7 @@ import RxSwift
 
 protocol ParameterModelProtocol {
     var userParameter: Observable<UserParameter> { get }
-    var gainedExp: Observable<Int> { get }
+    var gainedExp: Observable<Int64> { get }
 }
 
 final class ParameterModelImpl: ParameterModelProtocol {
@@ -11,7 +11,7 @@ final class ParameterModelImpl: ParameterModelProtocol {
         return Observable.just(UserParameter(fireAttack: 5, waterAttack: 5, windAttack: 5, soilAttack: 5, lightAttack: 5, darknessAttack: 10, maxHp: 100, defenseType: []))
     }
     
-    var gainedExp: Observable<Int> {
+    var gainedExp: Observable<Int64> {
         return Observable.just(100)
     }
 }
